requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==2.10
numpy==1.21.4
pandas==1.3.4
pika==1.2.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
